% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_file_size.R
\name{get_all_file_size}
\alias{get_all_file_size}
\title{get_all_file_size}
\usage{
get_all_file_size(path)
}
\arguments{
\item{path}{un chemin}
}
\value{
un tableau
}
\description{
Elle prend un chemin de dossier en paramètre, et map la fonction get_one_file_size pour n’obtenir qu’un seul tableau.
}
\examples{
get_all_file_size(path = ".")
}
