% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_file_size.R
\name{get_one_file_size}
\alias{get_one_file_size}
\title{get_one_file_size}
\usage{
get_one_file_size(file)
}
\arguments{
\item{file}{a file}
}
\value{
a dataframe with filename, size and extension
}
\description{
get_one_file_size
}
\examples{
temp_file_1 <- tempfile()
writeLines("Test content", temp_file_1)
get_one_file_size(temp_file_1)
}
